requirements.txt
flask==2.0.1
pyyaml==6.0
docker==5.0.3
requests==2.26.0
werkzeug==2.0.1 